# swagger.yml
openapi: 3.0.0
servers:
  - url: http://localhost:3000/api/v1
    description: Development Server
    # url: http://vault-backend.eba-wrntzw3k.ap-south-1.elasticbeanstalk.com/api/v1
    # description: AWS Elastic Bean Stalk
info:
  version: 1.0.0
  title: DevCamper API
  description: Backend API for the DevCamper application to manage bootcamps, courses, reviews, users and authentication
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  contact:
    name: 'Manoj Kumar G'
    url: 'https://www.linkedin.com/in/manojg381997/'
consumes: application/json
produces: application/json
paths:
...
  /courses:
    get:
      parameters:
        - in: query
          name: select
          example: 'title,description'
          schema:
            type: string
          required: false
          description: Select specific fields
        - in: query
          name: sort
          example: 'title'
          schema:
            type: string
          required: false
          description: Sort based on value
        - in: query
          name: page
          example: 1
          schema:
            type: number
          required: false
          description: Select page number
        - in: query
          name: limit
          example: 4
          schema:
            type: number
          required: false
          description: Select page limit
      summary: Get all Courses
      tags:
        - Courses
      responses:
        '200':
          description: On success
        '400':
          description: On error
  /bootcamps/{bootcampId}/courses:
    get:
      parameters:
        - in: path
          name: bootcampId
          example: '5d713995b721c3bb38c1f5d0'
          schema:
            type: string
          required: true
          description: Bootcamp ID
      summary: Get courses for specific bootcamp
      tags:
        - Courses
      responses:
        '200':
          description: On success
        '400':
          description: On error
    post:
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: bootcampId
          example: '5d713995b721c3bb38c1f5d0'
          schema:
            type: string
          required: true
          description: Bootcamp ID
      summary: Create course for specific bootcamp
      tags:
        - Courses
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#components/schemas/CreateCourseModel'
      responses:
        '200':
          description: On success
        '400':
          description: On error
...
components:
...
  securitySchemes:
    BearerAuth:
      type: apiKey
      in: header
      name: authorization